+++ src/collective/contact/facetednav/browser/actions/base.py
@@ -4,10 +4,12 @@ from plone.app.layout.viewlets.common import ViewletBase
 
 class BatchActionBase(ViewletBase):
     index = ViewPageTemplateFile('batchaction.pt')
-    klass = 'context'
-    onclick = None
-    name = None
-    weight = 1000
+    klass = 'context' # css class
+    onclick = None # onclick action. Must be set.
+    name = None # action id. Must be set.
+    weight = 1000 # info for actions order
+    # if True the action is activated when several elements are selected
+    multiple_selection = False
 
     def available(self):
         return True
@@ -15,11 +17,11 @@ class BatchActionBase(ViewletBase):
 
 class ActionBase(ViewletBase):
     index = ViewPageTemplateFile('action.pt')
-    klass = None
-    onclick = None
-    icon = None
-    name = None
-    weight = 1000
+    klass = None # css class
+    onclick = None # onclick action. Must be set.
+    icon = None # action icon. Must be set.
+    name = None # action id. Must be set.
+    weight = 1000 # info for actions order
 
     def available(self):
         return True